+++ modules/elementary_functions/src/cpp/NotEquals.cpp
@@ -88,6 +88,10 @@ ArrayOf
 NotEquals(ArrayOf& A, ArrayOf& B, bool& needToOverload)
 {
     needToOverload = false;
+    if (A.isReferenceType() || B.isReferenceType()) {
+        needToOverload = true;
+        return ArrayOf();
+	}
     VectorCheck(A, B, "~=");
     Class classCommon = FindCommonType(A, B, false);
     A.promoteType(classCommon);