+++ modules/elementary_functions/src/cpp/Transpose.cpp
@@ -27,7 +27,7 @@ Transpose(const ArrayOf& A, bool& needToOverload)
 {
     needToOverload = false;
     Class classA = A.getDataClass();
-    if ((classA < NLS_LOGICAL || A.isSparse()) && !(A.isCell() || A.isStruct())) {
+    if ((classA < NLS_LOGICAL || A.isSparse()) && !(A.isCell() || A.isStruct() || A.isStringArray())) {
         needToOverload = true;
         return ArrayOf();
     }